- added safe padding - enabled edge to edge - renamed package to wandering ball
