+++ composeApp/src/androidMain/kotlin/eu/heha/wanderingball/WanderingBallActivity.kt
@@ -1,23 +1,18 @@
-package eu.heha.meditation
+package eu.heha.wanderingball
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
+import androidx.activity.enableEdgeToEdge
 
-class MainActivity : ComponentActivity() {
+class WanderingBallActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
+        enableEdgeToEdge()
 
         setContent {
             App()
         }
     }
-}
-
-@Preview
-@Composable
-fun AppAndroidPreview() {
-    App()
 }
