+++ composeApp/src/commonMain/kotlin/eu/heha/wanderingball/ui/BounceRoute.kt
@@ -1,4 +1,4 @@
-package eu.heha.meditation.ui
+package eu.heha.wanderingball.ui
 
 import androidx.compose.runtime.Composable
 import androidx.lifecycle.viewmodel.compose.viewModel
