+++ composeApp/src/commonMain/kotlin/eu/heha/wanderingball/ui/ColorValue.kt
@@ -1,4 +1,4 @@
-package eu.heha.meditation.ui
+package eu.heha.wanderingball.ui
 
 import androidx.compose.material3.ColorScheme
 import androidx.compose.material3.MaterialTheme
