+++ composeApp/src/wasmJsMain/kotlin/eu/heha/wanderingball/main.kt
@@ -1,4 +1,4 @@
-package eu.heha.meditation
+package eu.heha.wanderingball
 
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.window.ComposeViewport
